Describe morphology using summary functions

Describe morphology using summary functions

describe(
  .list,
  what = mean,
  of = "neighbours",
  in_function_of = switch(of, neighbours = "distance", distance = "neighbours",
    pathlength = "distance"),
  at = NULL,
  connected = NULL,
  cumulative = FALSE,
  name = NULL
)

Arguments

.list

List with the distance matrix and within, and potentially a connected matrix, typically output from look_for(). Alternatively, this can be a data frame resulting from a describe() including a name argument, to add additional described summaries.

what

Summary function.

of

Quantity to summarise, and return as y. Can be "neighbours" (default), "distance" or "pathlength".

in_function_of

Quantity to return as x. The default is to use "distance" in case of is "neighbours" or "pathlength", and "neighbours" otherwise.

at

Distances at which to describe the results. Defaults to the integer sequence from 0L to the maximum search distance, if available, and otherwise the maximum distance found.

connected

Logical. Count connected or disconnected neighbours only. Defaults to NULL, for which all neighbours are considered.

cumulative

Logical. Provide cumulative results (distances <= threshold) or binned results (distances at center of provided breaks; default).

name

Name to use for the current described dataset in an extra name column. If provided, the input list is included as attribute to the output, and describe() can be repeated multiple times in a pipechain.

Value

Data frame with distance and neighbours columns. If name is provided, an additional name column is included for identifying different described summaries. In case a name column is there, the object has an extra morphology.list attribute, to enable piping into additional describe() calls.